Fix latest docs link
The graphql query has picked the latest version from the markdown pages,
the problem is now that a plugin was released which is newer as the
latest core. The latest doc page has now pointed to the version of the
plugin which is not available in the core. The fix is to skip all plugin
related docs to find the latest version of the core docs.