Fix caching for branch builds (#46)
The gatsby cache works at the moment only for the master branch. This is because gatsby writes the absolute path in its redux cache files and jenkins uses the branch name as part of the workspace path. With this change we execute the build always in the same directory of a docker container to get the same paths in the cache across all branches.