implement canonical links correctly (#53)
Currently, canonical tags are not created to the latest version of a document for a given plugin/the core but to the latest overall version for a plugin/the core (e.g. if a page /downloads was present in version 1.0 of a plugin but not in the next version 1.1, then the canonical link would incorrectly point to the non-existent 1.1 version). Additionally, the canonical link would point to latest instead of the actual version identifier and would also point to the documentation root and not to the actual page path. The latest version of a plugin/the core would not have any canonical link. This change fixes all of these issues.